improve user messages
GStreamer is super fast to start playing the mp3 file,
so there's no need for the "Loading Sound File..." message anymore.

And we needed a few more line-breaks in the "please send a pull request"
message printed to text terminal in the end.

Also, for some reason there's a JAVA exception still being printed
to the terminal and there's still a need for pressing CTRL-C to
actually stop audio playback and close the program :-P